+++ CP/core/in_sum_global_constraint.h
@@ -10,11 +10,11 @@
 class In_Sum_Global_Constraint: public Sum_Global_Constraint    /// This class implements the sum global constraint with condition in \ingroup core
 {
   protected:
-    int constant2;      ///< the upper bound value of the interval
+    long constant2;      ///< the upper bound value of the interval
   
 	public:
 		// constructors and destructor
-		In_Sum_Global_Constraint (vector<Variable *> & var, int cst1, int cst2);		///< constructs a new constraint which involves the variable in var and which checks whether the sum of the values of the variables in var in the interval [cst1, cst2]
+		In_Sum_Global_Constraint (vector<Variable *> & var, long cst1, long cst2);		///< constructs a new constraint which involves the variable in var and which checks whether the sum of the values of the variables in var in the interval [cst1, cst2]
 		In_Sum_Global_Constraint (In_Sum_Global_Constraint & c);	              	  ///< constructs a new constraint by copying the constraint c
 		
 		// basic functions
@@ -22,10 +22,10 @@ class In_Sum_Global_Constraint: public Sum_Global_Constraint    /// This class i
 		bool Revise (CSP * pb, unsigned int var, Support * ls, Deletion_Stack * ds) override;	 ///< returns true if the application of arc-consistency on the constraint w.r.t. the variable var deletes a value in the domain of var, false otherwise
     string Get_Relation () override;         ///< returns the relation used to compare the sum and the constant
     string Get_XCSP3_Expression() override;  ///< returns the string corresponding to the expression of the constraint in XCSP 3 format
-    int Get_Constant2 ();										 ///< returns the second constant to which the sum is compared
+    long Get_Constant2 ();										 ///< returns the second constant to which the sum is compared
     
   private:
-    bool Compare (int sum);                     ///< returns true if the sum satisfies the constraint, false otherwise
+    bool Compare (long sum);                     ///< returns true if the sum satisfies the constraint, false otherwise
 };
 
 
@@ -48,14 +48,14 @@ inline string In_Sum_Global_Constraint::Get_Relation ()
 }
 
 
-inline bool In_Sum_Global_Constraint::Compare (int sum)
+inline bool In_Sum_Global_Constraint::Compare (long sum)
 // returns true if the sum satisfies the constraint, false otherwise
 {
   return (constant <= sum) && (sum <= constant2);
 }
 
 
-inline int In_Sum_Global_Constraint::Get_Constant2 ()
+inline long In_Sum_Global_Constraint::Get_Constant2 ()
 // returns the second constant to which the sum is compared
 {
   return constant2;
